- tiny clean up in FSFactory; - save settings on return in options menu, saves also volume changes; - use size_t instead of unsigned to cast pointers (in asserts) to get it working on 64 bit linux; - disabled some standard libraries in lua, like io and os;

+++ source/shared_lib/sources/graphics/model.cpp
@@ -44,7 +44,7 @@ Vec3f* allocate_aligned_vec3_array(unsigned n) {
 }
 
 void free_aligned_vec3_array(Vec3f *ptr) {
-	assert(((unsigned)ptr & 0xF) == 0);
+	assert(((size_t)ptr & 0xF) == 0);
 	_mm_free(ptr);
 }
